X509 certificate related code in decdicated module
