+++ utils/manifest-utils/src/lib.rs
@@ -1,11 +1,12 @@
 pub mod golem_certificate;
+pub mod keystore;
 pub mod manifest;
 pub mod matching;
 pub mod policy;
 pub mod util;
 
 pub use manifest::*;
-pub use policy::{Keystore, Policy, PolicyConfig};
-pub use util::{
-    decode_data, DecodingError, KeystoreLoadResult, KeystoreManager, KeystoreRemoveResult,
-};
+// pub use keystore::
+pub use keystore::KeystoreManager;
+pub use policy::{Policy, PolicyConfig};
+pub use util::{decode_data, DecodingError};
